requirements.txt
python-telegram-bot==13.7
beautifulsoup4==4.12.2
requests==2.31.0
